+++ src/main/java/org/geekpower/common/enums/Deleted.java
@@ -1,18 +1,19 @@
 package org.geekpower.common.enums;
 
 public enum Deleted {
-    NO((short) 0), // 0，未删除
-    IS((short) 1), // 1，已删除（回收站）
-    REAL((short) 2),// 2，已彻底删除
+
+    NO((byte) 0), // 0，未删除
+    IS((byte) 1), // 1，已删除（回收站）
+    REAL((byte) 2),// 2，已彻底删除
     ;
 
-    private short code;
+    private byte code;
 
-    private Deleted(short code) {
+    private Deleted(byte code) {
         this.code = code;
     }
 
-    public short getCode() {
+    public byte getCode() {
         return code;
     }
