+++ src/main/java/org/geekpower/common/enums/MessageStatus.java
@@ -1,17 +1,18 @@
 package org.geekpower.common.enums;
 
 public enum MessageStatus {
-    DRAFT((short) 0), // 0，草稿，
-    FORMAL((short) 1),// 1，已生效
+
+    DRAFT((byte) 0), // 0，草稿，
+    FORMAL((byte) 1),// 1，已生效
     ;
 
-    private short code;
+    private byte code;
 
-    private MessageStatus(short code) {
+    private MessageStatus(byte code) {
         this.code = code;
     }
 
-    public short getCode() {
+    public byte getCode() {
         return code;
     }
